+++ spec/unit/heroku_run_spec.rb
@@ -51,13 +51,14 @@ describe "HerokuRun" do
 
       def run_obj.run_shell!
         @output = ""
-        @status = Object.new
+        `exit 1` # for $? on the next line
+        @status = $?
       end
 
       run_obj.call
 
       expect(run_obj.instance_variable_get(:@empty_fail_count)).to eq(3)
-      expect(stderr.string).to include("retrying the command.")
+      expect(stderr.string).to include("retrying...")
     end
 
     it "retries 0 times on NON empty result" do
@@ -66,7 +67,8 @@ describe "HerokuRun" do
 
       def run_obj.run_shell!
         @output = "not empty"
-        @status = Object.new
+        `exit 0` # for $? on the next line
+        @status = $?
       end
 
       run_obj.call
@@ -81,13 +83,14 @@ describe "HerokuRun" do
 
       def run_obj.run_shell!
         @output = ""
-        @status = Object.new
+        `exit 1` # for $? on the next line
+        @status = $?
       end
 
       run_obj.call
 
       expect(run_obj.instance_variable_get(:@empty_fail_count)).to eq(0)
-      expect(stderr.string).to_not include("retrying the command.")
+      expect(stderr.string).to_not include("retrying...")
     end
 
     it "retries 0 times on empty result when disabled via ENV var" do
@@ -99,13 +102,14 @@ describe "HerokuRun" do
 
         def run_obj.run_shell!
           @output = ""
-          @status = Object.new
+          `exit 1` # for $? on the next line
+          @status = $?
         end
 
         run_obj.call
 
         expect(run_obj.instance_variable_get(:@empty_fail_count)).to eq(0)
-        expect(stderr.string).to_not include("retrying the command.")
+        expect(stderr.string).to_not include("retrying...")
       ensure
         ENV["HATCHET_DISABLE_EMPTY_RUN_RETRY"] = original_env
       end