+++ lib/hatchet/heroku_run.rb
@@ -26,6 +26,7 @@ module Hatchet
       app: ,
       heroku: {},
       retry_on_empty: !ENV["HATCHET_DISABLE_EMPTY_RUN_RETRY"],
+      retry_delay: 0,
       raw: false,
       stderr: $stderr,
       timeout: 0)
@@ -35,6 +36,7 @@ module Hatchet
       @timeout = timeout
       @command = build_heroku_command(command, heroku || {})
       @retry_on_empty = retry_on_empty
+      @retry_delay = retry_delay
       @stderr = stderr
       @output = ""
       @status = nil
@@ -65,6 +67,7 @@ module Hatchet
           message << "\nreleases: #{@app.releases}"
           message << "\n#{caller.join("\n")}"
           @stderr.puts message
+          sleep(@retry_delay) # without run_multi, this will prevent occasional "can only run one free dyno" errors
           retry
         end
       rescue HerokuRunTimeoutError => e
@@ -77,6 +80,7 @@ module Hatchet
           message << "\nreleases: #{@app.releases}"
           message << "\n#{caller.join("\n")}"
           @stderr.puts message
+          sleep(@retry_delay) # without run_multi, this will prevent occasional "can only run one free dyno" errors
           retry
         end
       end