+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- Fix typo in the reaper `"Duplicate destroy attempted"` message (https://github.com/heroku/hatchet/pull/175)
+
 ## 7.3.3
 
 - Quiet personal tokens (https://github.com/heroku/hatchet/pull/148)