+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+- Set `init.defaultBranch` in ci:setup to suppress `git init` warning in Git 2.30+ (https://github.com/heroku/hatchet/issues/172)
 - Switch `heroku ci:install_heroku` to the Heroku CLI standalone installer rather than the APT install method (https://github.com/heroku/hatchet/issues/171)
 
 ## 7.3.3