+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## HEAD
 
 - Fix typo in the reaper `"Duplicate destroy attempted"` message (https://github.com/heroku/hatchet/pull/175)
+- Set `init.defaultBranch` in ci:setup to suppress `git init` warning in Git 2.30+ (https://github.com/heroku/hatchet/issues/172)
+- Switch `heroku ci:install_heroku` to the Heroku CLI standalone installer rather than the APT install method (https://github.com/heroku/hatchet/issues/171)
 
 ## 7.3.3
 
@@ -345,7 +347,6 @@
 - Add `rake hatchet:teardown_travis` task to put in `travis.yml`:
 
     after_script: bundle exec rake hatchet:teardown_travis
-
 
 ## 0.1.1