+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.?.?
 
 - Change HerokuRun#call to use exceptions for empty output retries in preparation of related work
+- Thread safety for HerokuRun (https://github.com/heroku/hatchet/issues/157)
 
 ## 7.3.4