+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change HerokuRun#call to use exceptions for empty output retries in preparation of related work
 - Thread safety for HerokuRun (https://github.com/heroku/hatchet/issues/157)
 - Timeout support for 'heroku run' (in case of networking, hanging or "server boot" tests etc), defaults to 60 s, global override `$HATCHET_DEFAULT_RUN_TIMEOUT`, per-test option `:timeout`
+- Delay for empty/timeout retries if run_multi is off, defaults to 1 s, global override `$HATCHET_RUN_RETRY_DELAY`
 
 ## 7.3.4